move program testing helpers into library
